Refactor settings dialogs into dedicated SettingsDialogs.kt
Moved ApiKeyTestDialog, ResetSettingsDialog, and AboutDialog to com.grindrplus.manager.ui.SettingsDialogs.kt.
Cleaned up SettingsUtils.kt to contain only business logic.
Cleaned up SettingsScreen.kt to use the new dialog location (implicit package access).
